chore(marketplace): add only node file
