+++ packages/marketplace/node-only/index.ts
@@ -1,0 +1,11 @@
+import dotenv from 'dotenv';
+
+dotenv.config();
+
+export const requireEnv = (env: string): string => {
+  const environment = process.env[env];
+  if (!environment) {
+    throw new Error(`${env} is not set`);
+  }
+  return environment;
+};
