+++ packages/marketplace/tsup.config.ts
@@ -2,5 +2,5 @@ import { extendConfig } from '@shared/tsup';
 
 export default extendConfig({
   dts: true,
-  entry: ['src/index.ts'],
+  entry: ['src/index.ts', '!scripts/deploy.ts', '!node-only/**/*.ts'],
 });
